This is the move  AbstractService for issue #136 This marks the old one as deprecated and provides a new interface/implementation in org.threadly.util
** force push to fix javadoc **
