+++ gradle.properties
@@ -1,2 +1,2 @@
 group = org.threadly
-version = 3.7.0
+version = 3.8.0-SNAPSHOT
