+++ pom.xml
@@ -2,7 +2,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>org.threadly</groupId>
   <artifactId>Threadly</artifactId>
-  <version>3.7.0</version>
+  <version>3.8.0-SNAPSHOT</version>
   <packaging>jar</packaging>
 
   <name>Threadly</name>
